Split get resources and get datasources
Split release.go and experimental.go so that datasources
and resources are in appropriately named files.

This also means we can avoid renaming imports.
